requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai==0.2.0
pydantic==2.0.0
markdown==3.5.1
weasyprint==60.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
gunicorn==21.2.0






